+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - *BREAKING CHANGE*: `PageRows` implements `Page<Row>` from `package:page`.
 
+- Query result pagination. 
+
 ## 0.0.1
 
 - Basic client.